Early LinkedGraph implementation.
Our current hypergraph suite does not provide a easy interface
for graph editing like we would need for something like fusion passes
or construction of a supergraph.

LinkedGraph is a implementation of a hypergraph, where
nodes and edges are seperately allocated, and reference counted.
It only works with DAGs, no need to check for cycles they should never
really happen in our computational graphs.

LinkedGraph is aggressivly optimized for graph editing, so
although concentually each node and edge can be thought of as
ref counted and allocated, we really avoid heap allocations in
most cases and instead pool most allocations.

addNode: O(1)
addEdge: O(|srcs|), basically O(1) because |srcs| is almost always less than 3.
removeNode: O(|N|) worst case, but just because we potentially have to
deallocate all nodes, if we remove the last referenced node.
In case we replace an edge section like
A -> B -> C with A -> C, this results in the minal overhead,
which is a pooled deallocation of B and the edges and a pooled
allocation of the new edge.
removeEdge: O(|N|) worst case, but again same argument as for removeNode.

The implementation is probably big time overkill, because all of the
runtimes are not amortized, so no vectors, we actually use doubly linked
list for stuff like incoming and outgoing edge lists, which may seem
weird, but the idea is that because we pool allocations anyway we still
get decent cache performance and iteration is far from the most
import operation for editing.

For the final implementation pass, where we build the supergraph,
we will most likely use a combination of a ConstGraph for blazily fast
iteration (over the ops applied to a tensor)
and a LinkedGraph for construction, maybe a AdjGraph also works we will
see.

+++ compiler/memory/hypergraph/AdjGraph.hpp
@@ -288,8 +288,8 @@ public:
     if (it == m_nodes.end()) {
       return {const_node_iterator{}, const_node_iterator{}};
     }
-    std::size_t idx = std::distance(m_nodes.begin(), it);
-    const_node_iterator begin{m_nodes.data() + idx, idx, m_nodes.size()};
+    std::ptrdiff_t idx = std::distance(m_nodes.begin(), it);
+    const_node_iterator begin{m_nodes.data() + idx, static_cast<std::size_t>(idx), m_nodes.size()};
     return {begin, const_node_iterator{m_nodes.data() + m_nodes.size(),
                                        m_nodes.size(), m_nodes.size()}};
   }
@@ -301,8 +301,8 @@ public:
     if (it == m_edges.end()) {
       return {const_edge_iterator{}, const_edge_iterator{}};
     }
-    std::size_t idx = std::distance(m_edges.begin(), it);
-    const_edge_iterator begin{m_edges.data() + idx, idx, m_edges.size()};
+    std::ptrdiff_t idx = std::distance(m_edges.begin(), it);
+    const_edge_iterator begin{m_edges.data() + idx, static_cast<std::size_t>(idx), m_edges.size()};
     return {begin, const_edge_iterator{m_edges.data() + m_edges.size(),
                                        m_edges.size(), m_edges.size()}};
   }
